requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
asgiref==3.7.2
    # via flask
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
blinker==1.6.2
    # via flask
build==0.10.0
    # via pip-tools
charset-normalizer==3.2.0
    # via aiohttp
click==8.1.6
    # via
    #   flask
    #   pip-tools
flask[async]==2.3.2
    # via
    #   -r requirements.in
    #   pytest-flask
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.1
    # via
    #   build
    #   pytest
pip-tools==7.3.0
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
pyproject-hooks==1.0.0
    # via build
pytest==7.4.0
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-flask
pytest-asyncio==0.21.1
    # via -r requirements.in
pytest-flask==1.2.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
werkzeug==2.3.6
    # via
    #   flask
    #   pytest-flask
wheel==0.41.1
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
